Make sure ntp service is running and watching conf
The state was installing ntp, but not making sure it runs. Furthermore I noticed that when I made changes to the used NTP servers, it wouldn't be used, because the service never watches/restarts. I tested that a few times and it fits my needs. I hope other folks will like it, too.